Move ad-hoc request/response records into api/dto/
Extracted inline record types from controllers and UserDataService into
a dedicated dto/ namespace (budgetbud.Dto) with files grouped by domain:
budget_dto.cs, user_dto.cs, social_dto.cs.
