+++ api/user_controller.cs
@@ -18,6 +18,7 @@
  * The source is available at: https://github.com/ParadoxZero/budgetbud
  */
 
+using budgetbud.Dto;
 using budgetbud.Exceptions;
 using budgetbud.Models;
 using budgetbud.Services;
@@ -40,7 +41,6 @@ public class UserController : ControllerBase
     _identityService = identityService;
   }
 
-  public record GetUserDetailsResponse(string nickName, string id);
   [HttpGet("details")]
   public async Task<IActionResult> GetUserDetails()
   {
@@ -51,7 +51,6 @@ public class UserController : ControllerBase
     return Ok(userDetails);
   }
 
-  public record UpdateNickNameInput(string nickName);
   [HttpPost("nickname")]
   public async Task<IActionResult> UpdateUserNickName([FromBody] UpdateNickNameInput input)
   {
